% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_HMM.R
\name{backward_HMM}
\alias{backward_HMM}
\title{Compute the backward probabilities in hidden Markov model}
\usage{
backward_HMM(transProb, emisProb, y.lst)
}
\arguments{
\item{transProb}{transition probabilities of the hidden Markov model.}

\item{emisProb}{emission probabilities of the hidden Markov model.}

\item{y.lst}{a sequence of observed states.}
}
\value{
a matrix of the backward probabilities. The first dimension is the state and the second is time.
}
\description{
Compute the backward probabilities in hidden Markov model
}
