% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3_q3.R
\name{baum_welch}
\alias{baum_welch}
\title{Hidden Markov models}
\usage{
baum_welch(
  observations,
  initial_model,
  tolerance = 0.001,
  max_iterations = 1000,
  callback = NULL
)
}
\arguments{
\item{observations}{a vector of observed values.}

\item{initial_model}{an object of type "hmm"; see \code{hmm}. Contains the initial values of the HMM parameters to use for optimization.}

\item{tolerance}{used for the stopping criterion in optimization.}

\item{max_iterations}{the maximum number of iterations to allow during optimization. A warning will be displayed if convergence is not achieved by then.}

\item{callback}{an optional function. If a subset of the parameters' true values are known before optimization time, the callback can be used to specify the true values and prevent optimization of those parameters. The callback must be a function of an "hmm" object and return an "hmm" object with the known parameter values replaced.}
}
\value{
an object of type "hmm" with the locally optimal parameter values.
}
\description{
\code{baum_welch} runs the Baum-Welch algorithm for HMM parameter inference.
}
\examples{
model <- hmm(
    initial_dist     = c(1/2, 1/2),
    transition_probs = rbind(c(0.98, 0.02), c(0.05, 0.95)),
    emission_probs   = rbind(rep(1/6, 6), c(1/10, 1/10, 1/2, 1/10, 1/10, 1/10)),
    states           = c("A", "B")
)

observations <- sample_hmm(100, model)$observations

initial_model <- hmm(
     initial_dist     = c(1/4, 3/4),
     transition_probs = rbind(c(1/2, 1/2), c(1/2, 1/2)),
     emission_probs   = rbind(rep(1/6, 6), rep(1/6, 6)),
     states           = c("A", "B")
 )

baum_welch(observations, initial_model)

# Assume we know the true initial distribution:
baum_welch(observations, initial_model,
    callback = \(m) { m$v <- model$v; m }
)

}
