% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baum_welch_HMM.R
\name{baum_welch_HMM}
\alias{baum_welch_HMM}
\title{Estimate the parameters in hidden Markov model using Baum-Welch algorithm}
\usage{
baum_welch_HMM(
  iniDist,
  transProb,
  emisProb,
  y.lst,
  max_iter = 100,
  delta = 1e-10
)
}
\arguments{
\item{iniDist}{initial distribution of the hidden Markov model.}

\item{transProb}{transition probabilities of the hidden Markov model.}

\item{emisProb}{emission probabilities of the hidden Markov model.}

\item{y.lst}{a sequence of observed states.}

\item{max_iter}{maximum number of iteration}

\item{delta}{threshold to test for convergence}
}
\value{
updated initial distribution, transition probabilities, and emission probabilities of HMM
}
\description{
Estimate the parameters in hidden Markov model using Baum-Welch algorithm
}
