% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3_q2.R
\name{blood_type_em}
\alias{blood_type_em}
\title{Blood genotype inference}
\usage{
blood_type_em(n, p_initial, tolerance = 0.001)
}
\arguments{
\item{n}{a named integer vector containing the observed counts of each blood type. Requires names are ("a", "ab", "b", "o").}

\item{p_initial}{a named numerical vector containing the initial values to use when optimizing for allele frequencies. Required names are ("a", "b", "o").}

\item{tolerance}{used for the stopping criterion in optimization.}
}
\value{
a named vector containing the estimated allele frequencies.
}
\description{
\code{blood_type_em} estimates population-level allele frequencies behind observed individuals' blood types using the EM algorithm. Specifically, we assume a Hardy-Weinberg equilibrium to relate the allele combinations (AA, BB, OO, AO, BO, AB) to allele frequencies (\eqn{p_A, p_B, p_O}).
}
\examples{
blood_type_em(
    n = c(a = 6, ab = 4, b = 55, o = 35),
    p_initial = c(a = 1 / 3, b = 1 / 3, o = 1 / 3)
)

}
