% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3EM.R
\name{EM}
\alias{EM}
\title{ABO Blood inference by EM Algorithm}
\usage{
EM(n, p, epilson)
}
\arguments{
\item{n}{different blood type counts}

\item{p}{nitial values of the allele frequency probabilities}

\item{epilson}{stopping criteria}
}
\value{
beta parameters,corresponding asymptotic confidence intervals, and a vector of the log-likelihoods from Newton_Raphson
}
\description{
ABO Blood inference by EM Algorithm
}
\examples{
n<-c(2,4,50,1)
p<-c(1/3,1/3,1/3)
EM(n,p,epilson=1e-8)
}
