% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_max.R
\name{exp_max}
\alias{exp_max}
\title{Expectation-Maximization algorithm for the ABO blood type problem}
\usage{
exp_max(nA, nAB, nB, nO, pA, pB)
}
\arguments{
\item{nA}{observed number of individuals with phenotype A.}

\item{nAB}{observed number of individuals with phenotype AB.}

\item{nB}{observed number of individuals with phenotype B.}

\item{nO}{observed number of individuals with phenotype O.}

\item{pA}{initial value of allele frequency probability of A.}

\item{pB}{initial value of allele frequency probability of B.}
}
\value{
allele frequency estimates of A, B, O
}
\description{
Expectation-Maximization algorithm for the ABO blood type problem
}
