% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_HMM.R
\name{forward_HMM}
\alias{forward_HMM}
\title{Compute the forward probabilities in hidden Markov model}
\usage{
forward_HMM(iniDist, transProb, emisProb, y.lst)
}
\arguments{
\item{iniDist}{initial distribution of the hidden Markov model.}

\item{transProb}{transition probabilities of the hidden Markov model.}

\item{emisProb}{emission probabilities of the hidden Markov model.}

\item{y.lst}{a sequence of observed states.}
}
\value{
a matrix of the forward probabilities. The first dimension is the state and the second is time.
}
\description{
Compute the forward probabilities in hidden Markov model
}
