% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient descent/steepest ascent method}
\usage{
gradient_descent(X, y, alpha, epilson, n.iter)
}
\arguments{
\item{X}{covariance matrix}

\item{alpha}{steep size}

\item{epilson}{stopping criteria}

\item{n.iter}{max iterartions}

\item{Y}{response vector}
}
\value{
beta parameters,corresponding asymptotic confidence intervals, and a vector of the log-likelihoods from Newton_Raphson
}
\description{
Gradient descent/steepest ascent method
}
\examples{
X<- matrix(x1,x2,..xn)
Y<-Y
gradient_descent(X,Y,1e-5,1e-8,1000)
}
