% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3_q3.R
\name{hmm}
\alias{hmm}
\title{Hidden Markov models}
\usage{
hmm(
  initial_dist,
  transition_probs,
  emission_probs,
  states = NULL,
  observed_values = NULL
)
}
\arguments{
\item{initial_dist}{a vector of initial probabilities for each hidden state.}

\item{transition_probs}{a matrix of probabilities for transitioning between hidden states. The (i,j)th element corresponds to P(j|i).}

\item{emission_probs}{a matrix of probabilities for emitting obserbed values at hidden states. The (i,j)th element corresponds to E(j|i).}

\item{states}{an optional vector of names for the hidden states. This is used to populate the column and row names of the parameter vectors appropriately.}

\item{observed_values}{an optional vector of names for the possible observed values. This is used to populate the column and row names of the parameter vectors appropriately.}
}
\value{
an object of class "hmm" for use with functions in this package.
}
\description{
\code{hmm} collects the parameters of a hidden Markov model for use with HMM-related functions in this package, including \code{sample_hmm} and \code{baum_welch}.
}
\examples{
model <- hmm(
    initial_dist     = c(1/2, 1/2),
    transition_probs = rbind(c(0.98, 0.02), c(0.05, 0.95)),
    emission_probs   = rbind(rep(1/6, 6), c(1/10, 1/10, 1/2, 1/10, 1/10, 1/10)),
    states           = c("A", "B")
)

}
