% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3_q3.R
\name{hmm_forward}
\alias{hmm_forward}
\title{Hidden Markov models}
\usage{
hmm_forward(model, observations)
}
\arguments{
\item{model}{an object of type "hmm"; see \code{hmm}.}

\item{observations}{a vector of observed values.}
}
\value{
the matrix \eqn{A} in the context of the algorithm, with number of rows equal to \code{length(observations)} and number of columns equal to the number of hidden states.
}
\description{
\code{hmm_forward} runs the HMM forward algorithm for likelihood computation.
}
\examples{
model <- hmm(
    initial_dist     = c(1/2, 1/2),
    transition_probs = rbind(c(0.98, 0.02), c(0.05, 0.95)),
    emission_probs   = rbind(rep(1/6, 6), c(1/10, 1/10, 1/2, 1/10, 1/10, 1/10)),
    states           = c("A", "B")
)

observations <- sample_hmm(100, model)$observations

hmm_forward(model, observations)

}
