% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{lm_qr}
\alias{lm_qr}
\title{Fitting linear models using a QR decomposition}
\usage{
lm_qr(formula, data)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted. See documentation for \code{lm} for details.}

\item{data}{a data frame containing the response and covariates in the model.}
}
\value{
a named vector of intercept and coefficient estimates.
}
\description{
\code{lm_qr} obtains OLS estimates of a linear regression model using the QR decomposition of the design matrix, \eqn{X = QR}.
The estimates solve the equation \eqn{R \beta = Q'y}.
}
\examples{
lm_qr(mpg ~ wt + hp, data = mtcars)

}
