% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{lm_svd}
\alias{lm_svd}
\title{Fitting linear models using a singular value decomposition}
\usage{
lm_svd(formula, data)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted. See documentation for \code{lm} for details.}

\item{data}{a data frame containing the response and covariates in the model.}
}
\value{
a named vector of intercept and coefficient estimates.
}
\description{
\code{lm_qr} obtains OLS estimates of a linear regression model using the SVD of the design matrix, \eqn{X = UDV'}.
The estimates solve the equation \eqn{V' \beta = D^{-1} U'y}.
}
\examples{
lm_svd(mpg ~ wt + hp, data = mtcars)

}
