% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.reg.R
\name{logistic.reg}
\alias{logistic.reg}
\title{Logistic regression using gradient descent or Newton-Raphson method}
\usage{
logistic.reg(x, y, threshold, max_iter, method, alpha = 0.1)
}
\arguments{
\item{x}{design matrix of predictor variables(including the intercept term).}

\item{y}{vector of binary response variable.}

\item{threshold}{threshold value to test for convergence.}

\item{max_iter}{maximum number of iterations.}

\item{method}{indicator of method; if method=1, use gradient descent; if method=2, use Newton-Raphson method.}

\item{alpha}{learning rate for gradient descent.}
}
\value{
MLEs of regression coefficients, their corresponding asymptotic confidence intervals, and a vector of the log-likelihoods “visited” by the chosen optimization option.
}
\description{
Logistic regression using gradient descent or Newton-Raphson method
}
