% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3_q1.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Fitting logistic regression models}
\usage{
logistic_regression(
  formula,
  data,
  method = "IRLS",
  tolerance = 1e-04,
  step_size = 0.001,
  max_iterations = 5000,
  rescale = F
)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted. See documentation for \code{lm} for details.}

\item{data}{data frame containing the response and covariates in the model.}

\item{method}{one of {"IRLS", "gradient"}, to obtain the MLEs using IRLS or gradient ascent.}

\item{tolerance}{used for the stopping criterion in optimization.}

\item{step_size}{used for the gradient ascent updates. Unused if \code{method = "IRLS"}.}

\item{max_iterations}{the maximum number of iterations to allow during optimization. A warning will be displayed if convergence is not achieved by then.}

\item{rescale}{logical; if \code{TRUE}, numeric columns in the data will be standardized before optimization and unstandardized after. This is for numerical stability, particularly when \code{method = "gradient"}. Coefficient and confidence interval outputs will still be in terms of the original scale.}
}
\value{
an object of class "logistic_regression". Supported S3 methods are \code{coef}, \code{confint}, and \code{plot} (for plotting log likelihoods over optimization time).
}
\description{
\code{logistic_regression} obtains maximum likelihood estimates of coefficients using either iteratively-reweighted least squares or gradient ascent.
}
\examples{
logistic_regression(I(mpg > 20) ~ disp, data = mtcars)

}
