% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_mult.R
\name{mat_mult}
\alias{mat_mult}
\title{Multiply two square matrices and a vector}
\usage{
mat_mult(a, b, x, opt)
}
\arguments{
\item{a}{a square matrix}

\item{b}{another square matrix}

\item{x}{a vector}

\item{opt}{argument indicating the order of multiplication}
}
\value{
The multiplication of a, b, and x
}
\description{
Multiply two square matrices and a vector
}
