% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework1.R
\name{matrix_prod}
\alias{matrix_prod}
\title{Multiply three matrices}
\usage{
matrix_prod(A, B, x, method = 1)
}
\arguments{
\item{A}{a M x N matrix}

\item{B}{a N x P matrix}

\item{x}{a P x 1 vector/matrix}

\item{method}{specify the order of multiplication}
}
\value{
If method=1, returns (AB)x, otherwise returns A(Bx)
}
\description{
Multiply three matrices
}
\examples{
matrix_prod(A, B, x, 1))
matrix_prod(A, B, x, 2))
}
