% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{multiply_matrices_and_vector}
\alias{multiply_matrices_and_vector}
\title{Multiply two square matrices and a vector}
\usage{
multiply_matrices_and_vector(A, B, x, matrices_first = T)
}
\arguments{
\item{A, B}{The matrices to multiply.}

\item{x}{The vector to multiply.}

\item{matrices_first}{If \code{TRUE}, evaluates \eqn{AB} first, otherwise evaluates \eqn{Bx} first. Useful for benchmarking the performance of the two multiplication algorithms.}
}
\value{
The product \eqn{ABx}.
}
\description{
\code{multiply_matrices_and_vector} returns the product of two square matrices and a vector.
The product can be computed in two ways, either by first multiplying the matrices, or by first
multiplying the second matrix by the vector.
}
\examples{
A <- matrix(1:9,   nrow = 3)
B <- matrix(11:19, nrow = 3)
x <- 1:3

multiply_matrices_and_vector(A, B, x)

}
