% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_chol_sim.R
\name{mvn_chol_sim}
\alias{mvn_chol_sim}
\title{Simulate from a multivariate normal distribution using Cholesky decomposition}
\usage{
mvn_chol_sim(n, mu, sigma)
}
\arguments{
\item{n}{the number of realizations.}

\item{mu}{a vector giving the means of the variables.}

\item{sigma}{a positive definite symmetric matrix specifying the covariance matrix of the variables.}
}
\value{
n realizations from the specified multivariate normal distribution.
}
\description{
Simulate from a multivariate normal distribution using Cholesky decomposition
}
\examples{
mu <- c(1,2)
sigma <- matrix(c(10,3,3,2),2,2)
mvn_chol_sim(10,mu,sigma)
}
