% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2mvn_cholesky.R
\name{mvn_cholesky}
\alias{mvn_cholesky}
\title{Sampling from MVN using Cholesky decoposition}
\usage{
mvn_cholesky(N, mu, sigma)
}
\arguments{
\item{N, }{number of samples}

\item{mu}{mean vector}

\item{sigma}{covariance matrix}
}
\value{
N realizations from MVN(mu,sigma)
}
\description{
Sampling from MVN using Cholesky decoposition
}
\examples{
n<-10
N<-100
mu<- mu<-matrix(rnorm(n),nrow=n,ncol=1)
sigma<- matrix(rWishart(1,n,diag(n)),nrow=n,ncol=n)
mvn_cholesky(N,mu,sigma)
}
