% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3Newton_Raphson.R
\name{Newton_Raphson}
\alias{Newton_Raphson}
\title{Newton-Raphson/Fisher scoring/IRLS method}
\usage{
Newton_Raphson(X, y, epilson, n.iter)
}
\arguments{
\item{X}{covariance matrix}

\item{epilson}{stopping criteria}

\item{n.iter}{max iterations}

\item{Y}{response vector}
}
\value{
beta parameters,corresponding asymptotic confidence intervals, and a vector of the log-likelihoods from Newton_Raphson
}
\description{
Newton-Raphson/Fisher scoring/IRLS method
}
\examples{
X<- matrix(x1,x2,..xn)
Y<-Y
Newton_Raphson(X,Y,1e-8,1000)
}
