% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_qr.R
\name{ols_qr}
\alias{ols_qr}
\title{OLS estimation of regression coefficients via QR decomposition}
\usage{
ols_qr(x, y)
}
\arguments{
\item{x}{design matrix of predictor varialbes}

\item{y}{vector of response variable}
}
\value{
Estimated coefficients of linear regression model
}
\description{
OLS estimation of regression coefficients via QR decomposition
}
