% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_svd.R
\name{ols_svd}
\alias{ols_svd}
\title{OLS estimation of regression coefficients via singular value decomposition}
\usage{
ols_svd(x, y)
}
\arguments{
\item{x}{design matrix of predictor variables}

\item{y}{vector of response variable}
}
\value{
Estimated coefficients of linear regression model
}
\description{
OLS estimation of regression coefficients via singular value decomposition
}
