% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2QLS_QR.R
\name{QLS_QR}
\alias{QLS_QR}
\title{QLS estimation using QR decomposition}
\usage{
QLS_QR(X, Y)
}
\arguments{
\item{X}{covariance matrix}

\item{Y}{response vector}
}
\value{
beta parameters from QLS estimation using QR decomposition
}
\description{
QLS estimation using QR decomposition
}
\examples{
X<- matrix(x1,x2,..xn)
Y<-Y
QLS_QR(X,Y)
}
