% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2QLS_SVD.R
\name{QLS_SVD}
\alias{QLS_SVD}
\title{QLS estimation using SVD decomposition}
\usage{
QLS_SVD(X, Y)
}
\arguments{
\item{X}{covariance matrix}

\item{Y}{response vector}
}
\value{
beta parameters from QLS estimation using QR decomposition
}
\description{
QLS estimation using SVD decomposition
}
\examples{
X<- matrix(x1,x2,..xn)
Y<-Y
QLS_SVD(X,Y)
}
