% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Random generation for the multivariate normal distribution}
\usage{
rmvnorm(N, mu, Sigma)
}
\arguments{
\item{N}{number of dimension \eqn{n} samples to generate.}

\item{mu}{mean vector, dimension \eqn{n}.}

\item{Sigma}{covariance matrix, dimension \eqn{n*n}.}
}
\value{
N realizations from \eqn{MVN(μ, Σ)}, as a dimension \eqn{N*n} matrix. Each row is an observation.
}
\description{
Given a dimension \eqn{n} mean vector and a dimension \eqn{n*n} covariance matrix, \code{rmnvorm} generates N realizations from the n-dimensional MVN distribution by transforming \eqn{N*n} iid standard normal samples, using the mean vector and the Cholesky decomposition of the covariance matrix.
}
\examples{
mu <- c(1, 5, 10, 50)
Sigma <- rbind(
    c(1.0,   0.5,  0.25, 0.125),
    c(0.5,   1.0,  0.5,  0.25 ),
    c(0.25,  0.5,  1.0,  0.5  ),
    c(0.125, 0.25, 0.5,  1.0  )
)

rmvnorm(100, mu, Sigma)

}
