% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3_q3.R
\name{sample_hmm}
\alias{sample_hmm}
\title{Hidden Markov models}
\usage{
sample_hmm(n, model)
}
\arguments{
\item{n}{the number of realizations to generate.}

\item{model}{an object of type "hmm"; see \code{hmm}.}
}
\value{
a tibble with columns for generated hidden states and observed values.
}
\description{
\code{sample_hmm} simulates realizations of hidden states and observed values from a hidden Markov model.
}
\examples{
model <- hmm(
    initial_dist     = c(1/2, 1/2),
    transition_probs = rbind(c(0.98, 0.02), c(0.05, 0.95)),
    emission_probs   = rbind(rep(1/6, 6), c(1/10, 1/10, 1/2, 1/10, 1/10, 1/10)),
    states           = c("A", "B")
)

sample_hmm(100, model)

}
