% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3sim_HMM.R
\name{sim_HMM}
\alias{sim_HMM}
\title{HMM simulation}
\usage{
sim_HMM(V, P, E, S, N)
}
\arguments{
\item{V}{initial distribution}

\item{P}{Transition Probability}

\item{E}{Emission Probability}

\item{S}{hidden states}

\item{N}{number of simulations}
}
\value{
N of simulated X's and Y's
}
\description{
HMM simulation
}
\examples{
S<-c("1","2")
P=matrix(c(0.98,0.05,0.02,0.95),nrow=2)
E=matrix(c(1/6,1/10,1/6,1/10,1/6,1/2,1/6,1/10,1/6,1/10,1/6,1/10,1/6,1/10),nrow=2)
V=matrix(c(1/2,1/2))

sim_HMM(V,P,E,S,N=100)
}
