% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_HMM.R
\name{simulate_HMM}
\alias{simulate_HMM}
\title{Simulate hidden and observed states for a hidden Markov model}
\usage{
simulate_HMM(iniDist, transProb, emisProb, hidden, observed, length, seed)
}
\arguments{
\item{iniDist}{initial distribution of the hidden Markov model}

\item{transProb}{transition probabilities of the hidden Markov model.}

\item{emisProb}{emission probabilities of the hidden Markov model.}

\item{hidden}{vector of hidden states.}

\item{observed}{vector of observed states.}

\item{length}{length of the simulated sequence of the hidden and observed states.}

\item{seed}{seed number to generate random number.}
}
\value{
a sequence of hidden and observed states.
}
\description{
Simulate hidden and observed states for a hidden Markov model
}
