% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats230-package.R
\docType{package}
\name{stats230-package}
\alias{stats230}
\alias{stats230-package}
\title{stats230: Implementations of Algorithms Discussed in UCI Stats 230}
\description{
Package for UCI Stats 230 course assignments, under Prof. Vladimir Minin.
}
\author{
\strong{Maintainer}: Thanasi Bakis \email{thanasibakis@gmail.com}

}
\keyword{internal}
