% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baum_welch_HMM.R
\name{update_param_HMM}
\alias{update_param_HMM}
\title{Update the parameters of the hidden Markov model}
\usage{
update_param_HMM(iniDist, transProb, emisProb, y.lst)
}
\arguments{
\item{iniDist}{initial distribution of the hidden Markov model.}

\item{transProb}{transition probabilities of the hidden Markov model.}

\item{emisProb}{emission probabilities of the hidden Markov model.}

\item{y.lst}{a sequence of observed states.}
}
\value{
updated initial distribution, transition probabilities, and emission probabilities of HMM
}
\description{
Update the parameters of the hidden Markov model
}
