% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3Baum_Welch.R
\name{Baum_Welch}
\alias{Baum_Welch}
\title{Baum Welch HMM}
\usage{
Baum_Welch(V0, P0, E0, S, obs, n.iter, epilson)
}
\arguments{
\item{S}{hidden states}

\item{obs}{simulated observations}

\item{V}{Random initial distribution}

\item{P}{Random Transition Probability}

\item{E}{Random Emission Probability}
}
\value{
estimated True initial distribution,Transition Probability and Random Emission Probability
}
\description{
Baum Welch HMM
}
\examples{
S<-c("1","2")
P=matrix(c(0.75,0.25,0.10,0.9),nrow=2)
E=matrix(c(1/6,1/6,1/6,1/6,1/6,1/6,1/6,1/6,1/6,1/6,1/6,1/6,1/6,1/6),nrow=2)
V=matrix(c(1/4,3/4))
obs=c(1,2,3,4,5,6,7)

Baum_Welch(V,P,E,S,obs)
}
