% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3forward_HMM.R
\name{forward_HMM}
\alias{forward_HMM}
\title{Forward HMM simulation}
\usage{
forward_HMM(V, P, E, S, obs)
}
\arguments{
\item{V}{initial distribution}

\item{P}{Transition Probability}

\item{E}{Emission Probability}

\item{S}{hidden states}

\item{obs}{simulated obersvations}
}
\value{
forwards probabilities
}
\description{
Forward HMM simulation
}
\examples{
S<-c("1","2")
P=matrix(c(0.98,0.05,0.02,0.95),nrow=2)
E=matrix(c(1/6,1/10,1/6,1/10,1/6,1/2,1/6,1/10,1/6,1/10,1/6,1/10,1/6,1/10),nrow=2)
V=matrix(c(1/2,1/2))
obs=c(1,2,3,4,5,6,7)

forward_HMM(V,P,E,S,obs)
}
